Resolves #21
In order to be able to render multiple charts of the same type on a
page, we check to see if the lava.charts.<type> object exists and if
not, initialise it.

```javascript
if ( typeof lava.charts.<type> == "undefined" ) { lava.charts.<type> = {}; }'
```

We then initialise the individual charts as an object, keyed by it's
label on the chart type

```javascript
lava.charts.<type>.<label> = {chart:null,draw:null,data:null,options:null,formats:[]};
```

In doing so, we preserve the rest of the functionality, without
overriding the chart types, which (I believe) can be attributed to
variables bubbling up i.e. the laste declared chart <type> of <label>
will override preceding instances.